move player color enum to core lib
